add-edit-endpoint-refactor
Add and edit endpoints are  merged together.

+++ MVCBlogApp/Controllers/PostPublishingController.cs
@@ -1,8 +1,10 @@
+using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using MVCBlogApp.Interface;
 using MVCBlogApp.Models;
 using MVCBlogApp.Models.DTO.Post;
 using MVCBlogApp.Models.ViewModel;
+using MVCBlogApp.Models.ViewModel.Post;
 
 namespace MVCBlogApp.Controllers;
 
@@ -10,11 +12,14 @@ public class PostPublishingController : Controller
 {
     private readonly IPostPublishingService  _postPublishingService;
     private readonly ICategoryService _categoryService;
+    private readonly IMapper _mapper;
     public PostPublishingController(IPostPublishingService postPublishingService,
-        ICategoryService categoryService)
+        ICategoryService categoryService,
+        IMapper mapper)
     {
         _postPublishingService = postPublishingService;
         _categoryService = categoryService;
+        _mapper = mapper;
     }
 
     [HttpGet]
@@ -24,44 +29,51 @@ public class PostPublishingController : Controller
     }
 
     [HttpGet]
-    public async Task<IActionResult> NewPost()
-    {
-        return View(new NewPostView { PostCategories = await _categoryService.GetCategories() });
+    [Route("PostPublishing/Edit")] 
+    public async Task<IActionResult> Edit(int? postId) {
+        if (postId is > 0) {
+            var post = await _postPublishingService.GetWholePostAsync(postId.Value);
+            if (post != null) {
+                return View(new EditPostView {
+                    PostCategories = await _categoryService.GetCategories(),
+                    EditedPostId = postId,
+                    PostDto = _mapper.Map<AddOrEditPostDto>(post)
+                });
+            }
+        }
+        return View(new EditPostView { PostCategories = await _categoryService.GetCategories() });
     }
-
+    
     [HttpPost]
-    public async Task<IActionResult> AddNewPostAsync(NewPostView postView)
+    [Route("PostPublishing/Save")] 
+    public async Task<IActionResult> Save([FromForm] EditPostView model)
     {
-        await _postPublishingService.AddNewPostAsync(postView.PostDto);
+        if (!ModelState.IsValid) {
+            model.PostCategories = await _categoryService.GetCategories();
+            return View("Edit", model);
+        }
+
+        if (model.EditedPostId.HasValue && model.EditedPostId > 0) {
+            await _postPublishingService.EditPostAsync(model.PostDto, model.EditedPostId.Value);
+            TempData["Message"] = "Post edited successfully";
+        } else {
+            await _postPublishingService.AddPostAsync(model.PostDto);
+            TempData["Message"] = "Post added successfully";
+        }
+    
         return RedirectToAction("Index");
     }
 
     [HttpDelete]
-    public async Task<IActionResult> DeletePostAsync(int postId)
+    [Route("PostPublishing/Delete")] 
+    public async Task<IActionResult> DeleteAsync(int postId)
     {
-        await _postPublishingService.DeletePostAsync(postId);
+        if (postId is > 0) {
+            await _postPublishingService.DeletePostAsync(postId);
+            TempData["Message"] = "Post deleted successfully";
+        } else {
+            TempData["Warning"] = "Post id is not valid";            
+        }
         return RedirectToAction("Index");
-    }
-    
-    [HttpGet("id")]
-    public async Task<IActionResult> EditPost(int postId)
-    {
-        var post = await _postPublishingService.GetWholePostAsync(postId);
-        var postToEdit = new AddOrEditPostDto
-        {
-            Title = post.Title,
-            Content = post.Content,
-            Slug = post.Slug,
-            IsPublished = post.IsPublished,
-            PostCategoryId = post.PostCategoryId
-        };
-        return View(postToEdit);
-    }
-    
-    [HttpPut("id")]
-    public async Task<IActionResult> EditPost(AddOrEditPostDto editPostDto,int postId)
-    {
-        var post = await _postPublishingService.EditPostAsync(editPostDto, postId);
-        return Ok(post);
     }
 }
